+++ Tests/Unit/Core/Model/BlankNodeIdentifierTest.php
@@ -1,0 +1,62 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the "schema" extension for TYPO3 CMS.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE.txt file that was distributed with this source code.
+ */
+
+namespace Brotkrueml\Schema\Tests\Unit\Core\Model;
+
+use Brotkrueml\Schema\Core\Model\BlankNodeIdentifier;
+use Brotkrueml\Schema\Core\Model\NodeIdentifierInterface;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * @runTestsInSeparateProcesses
+ */
+class BlankNodeIdentifierTest extends TestCase
+{
+    /**
+     * @test
+     */
+    public function subjectIsInstanceOfNodeIdentifierInterface(): void
+    {
+        self::assertInstanceOf(NodeIdentifierInterface::class, new BlankNodeIdentifier());
+    }
+
+    /**
+     * @test
+     */
+    public function getIdReturnsBlankIdentifier(): void
+    {
+        $subject = new BlankNodeIdentifier();
+
+        self::assertSame('_:b0', $subject->getId());
+    }
+
+    /**
+     * @test
+     */
+    public function getIdOnConsecutiveInstantiationsReturnsAscendingBlankIdentifiers(): void
+    {
+        $actual = [
+            (new BlankNodeIdentifier())->getId(),
+            (new BlankNodeIdentifier())->getId(),
+            (new BlankNodeIdentifier())->getId(),
+            (new BlankNodeIdentifier())->getId(),
+        ];
+
+        $expected = [
+            '_:b0',
+            '_:b1',
+            '_:b2',
+            '_:b3',
+        ];
+
+        self::assertSame($expected, $actual);
+    }
+}