+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Node identifier and blank node identifier (#65)
+
 ### Changed
 - Context moved from http://schema.org to https://schema.org/ (#58)
 - By default, markup is added to noindex pages, a configuration setting is available for deactivation (#60)